// src/common/interfaces/shop.interface.ts
export interface IShop {
  id: string;
  baseUrl: string;
  shopifyTokenApi: string;
  shopifyKeyApi: string;
  shopifySecretKeyApi: string;
  urlStore: string;
  shopName: string;
  status: string;
  hostVersion: string;
  createdAt: string;
  updatedAt: string;
  userId: string;
  shopAvatar: string;
}